use non-root users in Docker image For building it is neccessary to use non-root, because otherwise the prepack scripts for e.g. building the matrix-bot-sdk fork don't run.
For normal running it also makes sense to use a non-root user to limit
the permissions in case of a breach.

This commit adds the user via Dockerfile and switches to it
in the entrypoint via `su`. The latter enables us to chown /data,
in case it already contains files or otherwise has wrong permissions.