add category name to room name format vars
Some Discord guilds make use of "categories" to further subdivide the
channels within the guild. This allows administrators to include this
category name into the room display name. This is escpecially
usefull, when multiple channels in the same guild have the same name
and are only differentiated by their category.

The discord.js API seems to suggest that categories could be
arbitrarily nested (as categories have a `parent` property, too), but
I could not find a way to create such nested categories in the Discord
UI. This change only exposes the category directly containing the
channel.